generalized to not only use Roqua namespace
