+++ lib/opencpu/configuration.rb
@@ -1,0 +1,7 @@
+module OpenCPU
+  class Configuration
+    attr_accessor :endpoint_url
+    attr_accessor :username
+    attr_accessor :password
+  end
+end
