+++ lib/opencpu/version.rb
@@ -1,0 +1,3 @@
+module OpenCPU
+  VERSION = "0.3.0"
+end
