--- lib/roqua/opencpu/configuration.rb
@@ -1,9 +1,0 @@
-module Roqua
-  module OpenCPU
-    class Configuration
-      attr_accessor :endpoint_url
-      attr_accessor :username
-      attr_accessor :password
-    end
-  end
-end
