--- lib/roqua/opencpu/version.rb
@@ -1,5 +1,0 @@
-module Roqua
-  module OpenCPU
-    VERSION = "0.3.0"
-  end
-end
