--- spec/lib/roqua_spec.rb
@@ -1,1 +1,0 @@
-require File.expand_path('../spec_helper.rb', __dir__)
