+++ spec/spec_helper.rb
@@ -1,6 +1,6 @@
 require 'vcr'
 
-require File.expand_path('../lib/roqua/opencpu.rb', __dir__)
+require File.expand_path('../lib/opencpu.rb', __dir__)
 
 VCR.configure do |c|
   c.cassette_library_dir = 'spec/fixtures/vcr_cassettes'
